project_root/requirements.txt
tensorflow
pandas
numpy
pathlib
FastAPI
fastapi
uvicorn
scikit-learn
pillow
opencv-python-headless
jinja2
python-multipart
gdown